requirements.txt
# API framework
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.6

# Async HTTP client
aiohttp>=3.8.1

# Configuration & data
pydantic<2.0.0,>=1.10.7
python-dotenv>=0.21.0

# STT
openai-whisper>=20230314
ffmpeg-python>=0.2.0

# LLM & quantization
transformers>=4.30.0
torch>=2.0.0
bitsandbytes>=0.39.0
accelerate>=0.18.0

# RAG
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
numpy>=1.23.5
langchain>=0.0.178

# File processing
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.0.10
pandas>=1.5.2

# Banking adapters & utils
# (no extra lib beyond aiohttp & stdlib)

# Security
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4

# TTS
pyttsx3>=2.90
